pull code main

+++ src/main/java/com/example/casestudymodule4/service/IUserService.java
@@ -3,5 +3,5 @@ package com.example.casestudymodule4.service;
 import com.example.casestudymodule4.model.User;
 
 public interface IUserService {
-    User findById(Long id);
+    User findById(Integer id);
 }
